+++ src/main/java/com/yjp/delivery/service/ShopService.java
@@ -1,16 +1,12 @@
 package com.yjp.delivery.service;
 
 import com.yjp.delivery.common.validator.ShopValidator;
-import com.yjp.delivery.common.validator.UserValidator;
-import com.yjp.delivery.controller.shop.dto.request.ShopLikeReq;
 import com.yjp.delivery.controller.shop.dto.response.MenuGetRes;
 import com.yjp.delivery.controller.shop.dto.response.ShopGetAllRes;
 import com.yjp.delivery.controller.shop.dto.response.ShopGetRes;
-import com.yjp.delivery.controller.shop.dto.response.ShopLikeRes;
 import com.yjp.delivery.store.entity.MenuEntity;
 import com.yjp.delivery.store.entity.ShopEntity;
 import com.yjp.delivery.store.entity.ShopLikeEntity;
-import com.yjp.delivery.store.entity.UserEntity;
 import com.yjp.delivery.store.repository.ShopLikeRepository;
 import com.yjp.delivery.store.repository.ShopRepository;
 import com.yjp.delivery.store.repository.UserRepository;
@@ -47,56 +43,21 @@ public class ShopService {
         return ShopServiceMapper.INSTANCE.toShopGetRes(shopEntity);
     }
 
-    public ShopLikeRes likeShop(ShopLikeReq shopLikeReq) {
-        ShopEntity shopEntity = getShopEntity(shopLikeReq.getShopId());
-        UserEntity userEntity = getUserEntity(shopLikeReq.getUserId());
-
-        ShopLikeEntity shopLikeEntity = getShopLike(shopEntity, userEntity);
-        ShopValidator.checkAlreadyLiked(shopLikeEntity);
-
-        shopLikeRepository.save(ShopLikeEntity.builder()
-            .shopId(shopEntity)
-            .userId(userEntity)
-            .build());
-        return new ShopLikeRes();
-    }
-
-    public ShopLikeRes unLikeShop(ShopLikeReq shopLikeReq) {
-        ShopEntity shopEntity = getShopEntity(shopLikeReq.getShopId());
-        UserEntity userEntity = getUserEntity(shopLikeReq.getUserId());
-
-        ShopLikeEntity shopLikeEntity = getShopLike(shopEntity, userEntity);
-        ShopValidator.checkNotYetLiked(shopLikeEntity);
-
-        shopLikeRepository.delete(shopLikeEntity);
-        return new ShopLikeRes();
-    }
-
-    private ShopEntity getShopEntity(Long shopId) {
-        ShopEntity shopEntity = shopRepository.findByShopId(shopId);
-        ShopValidator.validate(shopEntity);
-        return shopEntity;
-    }
-
-    private UserEntity getUserEntity(Long userId) {
-        UserEntity userEntity = userRepository.findByUserId(userId);
-        UserValidator.validate(userEntity);
-        return userEntity;
-    }
-
-    private ShopLikeEntity getShopLike(ShopEntity shopEntity, UserEntity userEntity) {
-        return shopLikeRepository.findByShopIdAndUserId(shopEntity, userEntity);
-    }
-
     @Mapper
     public interface ShopServiceMapper {
 
         ShopServiceMapper INSTANCE = Mappers.getMapper(ShopServiceMapper.class);
 
+        @Mapping(source = "shopLikeEntities", target = "like")
+        default int toLike(List<ShopLikeEntity> shopLikeEntities) {
+            return shopLikeEntities.size();
+        }
+
         MenuGetRes toMenuGetRes(MenuEntity menuEntity);
 
         List<MenuGetRes> toMenuGetReses(List<MenuEntity> menuEntities);
 
+        @Mapping(source = "shopLikeEntities", target = "like")
         @Mapping(source = "menuEntities", target = "menuGetReses")
         ShopGetRes toShopGetRes(ShopEntity shopEntity);
